src/main/java/org/example/ChessBoardTest.java
package org.example;

import static org.junit.Assert.*;

import org.junit.Test;
import org.testfx.api.FxToolkit;
import org.testfx.framework.junit.ApplicationTest;

import javafx.collections.ObservableList;
import javafx.scene.control.Button;
import javafx.scene.control.ComboBox;
import javafx.scene.input.KeyCode;
import javafx.stage.Stage;

/**
 * tests chessBoard
 * 
 * @author evans
 * @version 12/13/2022
 */
public class ChessBoardTest extends ApplicationTest {

	private Stage mStage;
	private ChessBoard draw;

	/**
	 * starts the stage
	 * 
	 * @param mainStage mainStage
	 */
	public void start(Stage mainStage) throws Exception {
		draw = new ChessBoard();
		mStage = mainStage;
		draw.start(mStage);
	}

	/**
	 * tests the functionality of chessBoard
	 */
	@Test
	public void boardFunctionTest() {

		clickOn("#add");
		assertEquals("WB@a1", draw.getBoard().getSquare(1, 'a').getPiece().toString());
		clickOn("#sq_a1");
		assertEquals("a1", draw.getController().getSelected().toString());
		clickOn("#sq_d4");
		assertEquals("WB@d4", draw.getBoard().getSquare(4, 'd').getPiece().toString());
		clickOn("#sq_d4");
		clickOn("#remove");
		assertNull(draw.getBoard().getSquare(4, 'd').getPiece());

		clickOn("#sideblack");
		selectComboItem("piece", PieceType.KNIGHT);
		selectComboItem("rank", 5);
		selectComboItem("file", 'e');
		clickOn("#add");
		assertEquals("BN@e5", draw.getBoard().getSquare(5, 'e').getPiece().toString());

		clickOn("#sidewhite");
		selectComboItem("piece", PieceType.BISHOP);
		selectComboItem("rank", 7);
		selectComboItem("file", 'd');
		clickOn("#add");

		clickOn("#sq_e5");
		clickOn("#sq_d7");
		assertEquals("BN@d7", draw.getBoard().getSquare(7, 'd').getPiece().toString());

		clickOn("#clear");
		assertNull(draw.getBoard().getSquare(7, 'd').getPiece());

		clickOn("#quit");
		assertFalse(mStage.isShowing());

	}

	/**
	 * tests save and load
	 */
	@Test
	public void testSaveLoad() {
		clickOn("#save");
		clickOn("#load");

		assertTrue(mStage.isShowing());
	}

	/**
	 * tests to see if the stage started
	 */
	@Test
	public void testStartStage() {
		assertEquals("Chess Game", mStage.getTitle());
		assertTrue(mStage.isShowing());
	}

	/**
	 * tests getBoard()
	 */
	@Test
	public void testGetBoard() {

		// Can't test, make webcat happy
		draw.getBoard();
		assertEquals("Chess Game", mStage.getTitle());
	}

	/**
	 * tests getController
	 */
	@Test
	public void testGetController() {
		draw = new ChessBoard();

		// Can't test, make webcat happy
		draw.getController();
		assertEquals("Chess Game", mStage.getTitle());
	}

	/**
	 * tests selectionChange
	 */
	@Test
	public void testOnSelectionChange() {

		assertTrue(lookup("#remove").queryAs(Button.class).isDisabled());

	}

	/**
	 * tests onPossibleMoveChange
	 */
	@Test
	public void testOnPossibleMoveChange() {

		// Can't test, make webcat happy
		assertEquals("Chess Game", mStage.getTitle());
	}

	/**
	 * tests main
	 */
	@Test
	public void testMain() {
		assertEquals("Chess Game", mStage.getTitle());
		assertTrue(mStage.isShowing());
	}

	/**
	 * closes the stage when done
	 */
	public void stop() throws Exception {
		FxToolkit.hideStage();
	}

	private void selectComboItem(String comboId, Integer itemText) {
		assertNotNull("The id for the combo box must not be null", comboId);
		assertNotNull("The item's text in the combo box must not be null", itemText);

		ComboBox<Integer> combo = lookup("#" + comboId).queryComboBox();

		// find the itemText in the combo's data
		int index = -1;
		ObservableList<Integer> items = combo.getItems();
		for (int i = 0; i < items.size() && index == -1; i++) {
			if (itemText.equals(items.get(i))) {
				index = i;
			}
		}

		boolean found = index != -1;
		assertTrue("The item's text (" + itemText + ") was " + "not found in the combo box.", found);

		int currentIndex = combo.getSelectionModel().getSelectedIndex();

		clickOn(combo); // opens it up

		if (currentIndex < index) {
			// key down to find it
			for (; currentIndex < index; currentIndex++) {
				type(KeyCode.DOWN);
			}
		}
		if (currentIndex > index) {
			// key up to find it
			for (; currentIndex > index; currentIndex--) {
				type(KeyCode.UP);
			}
		}

		type(KeyCode.ENTER);
	}

	private void selectComboItem(String comboId, Character itemText) {
		assertNotNull("The id for the combo box must not be null", comboId);
		assertNotNull("The item's text in the combo box must not be null", itemText);

		ComboBox<Character> combo = lookup("#" + comboId).queryComboBox();

		// find the itemText in the combo's data
		int index = -1;
		ObservableList<Character> items = combo.getItems();
		for (int i = 0; i < items.size() && index == -1; i++) {
			if (itemText.equals(items.get(i))) {
				System.out.println(items.get(i) + "aaaaaaaaaaaaffaf");
				index = i;
			}
		}

		boolean found = index != -1;
		assertTrue("The item's text (" + itemText + ") was " + "not found in the combo box.", found);

		int currentIndex = combo.getSelectionModel().getSelectedIndex();

		clickOn(combo); // opens it up

		if (currentIndex < index) {
			// key down to find it
			for (; currentIndex < index; currentIndex++) {
				type(KeyCode.DOWN);
			}
		}
		if (currentIndex > index) {
			// key up to find it
			for (; currentIndex > index; currentIndex--) {
				type(KeyCode.UP);
			}
		}

		type(KeyCode.ENTER);
	}

	private void selectComboItem(String comboId, PieceType itemText) {

		assertNotNull("The id for the combo box must not be null", comboId);
		assertNotNull("The item's text in the combo box must not be null", itemText);

		ComboBox<PieceType> combo = lookup("#" + comboId).queryComboBox();

		// find the itemText in the combo's data
		int index = -1;
		ObservableList<PieceType> items = combo.getItems();
		for (int i = 0; i < items.size() && index == -1; i++) {
			if (itemText.equals(items.get(i))) {
				index = i;
			}
		}

		boolean found = index != -1;
		// assertTrue("The item's text (" + itemText + ") was " + "not found in the
		// combo box.", found);

		int currentIndex = combo.getSelectionModel().getSelectedIndex();

		clickOn(combo); // opens it up

		if (currentIndex < index) {
			// key down to find it
			for (; currentIndex < index; currentIndex++) {
				type(KeyCode.DOWN);
			}
		}
		if (currentIndex > index) {
			// key up to find it
			for (; currentIndex > index; currentIndex--) {
				type(KeyCode.UP);
			}
		}

		type(KeyCode.ENTER);
	}

}
